Hierarchical Dirichlet Process Model

Hierarchical Dirichlet process (HDP) is a powerful mixed-membership model for the unsupervised analysis of grouped data. Unlike its finite counterpart, latent Dirichlet allocation, the HDP topic model infers the number of topics from the data. Here we have used Online HDP, which provides the speed of online variational Bayes with the modeling flexibility of the HDP. The idea behind Online variational Bayes in general is to optimize the variational objective function with stochastic optimization.The challenge we face is that the existing coordinate ascent variational Bayes algorithms for the HDP require complicated approximation methods or numerical optimization. This model utilises stick breaking construction of Hdp which enables it to allow for coordinate-ascent variational Bayes without numerical approximation.

model_hdp(corpus, id2word, ...)

# S3 method for mm_file
model_hdp(corpus, id2word, ...)

# S3 method for mm
model_hdp(corpus, id2word, ...)

# S3 method for python.builtin.list
model_hdp(corpus, id2word, ...)

# S3 method for list
model_hdp(corpus, id2word, ...)

Arguments

corpus
Model as returned by mmcorpus_serialize.

id2word
Dictionary for the input corpus, as returned by corpora_dictionary.

...
Any other options, from the official documentation.

Details

This is a non-parametric bayesian method: notice the lack of num_topics argument.